+++ node/src/types/kernel/api/snippet/getSnippet/response.d.ts
@@ -57,7 +57,7 @@ export interface ISnippet {
     /**
      * snippet memo
      */
-    memo: string;
+    memo?: string;
     /**
      * snippet name
      */